perbaiki hak akses untuk kadis

+++ admin/mutasi/del_mutasi.php
@@ -1,4 +1,12 @@
 <?php
+if($_SESSION['ses_level'] == "Kadis"){
+    echo "<script>
+    Swal.fire({title: 'Anda tidak punya akses ke menu ini!',text: '',icon: 'error',confirmButtonText: 'OK'
+    }).then((result) => {if (result.value){
+        window.location = 'index.php?page=data-mutasi';
+        }
+    })</script>";
+}else{
     $sql_hapus = "DELETE FROM data_mutasi WHERE id_mutasi='".$_GET['kode']."'";
     $query_hapus = mysqli_query($koneksi, $sql_hapus);
     if ($query_hapus) {
@@ -12,3 +20,4 @@
         }).then((result) => {if (result.value) {window.location = 'index.php?page=data-mutasi'
         ;}})</script>";
     }
+}
