+++ admin/pegawai/del_pegawai.php
@@ -8,16 +8,25 @@ if (isset($_GET['kode'])) {
 ?>
 
 <?php
-$sql_hapus = "DELETE FROM data_pegawai WHERE id_pegawai='" . $_GET['kode'] . "'";
-$query_hapus = mysqli_query($koneksi, $sql_hapus);
-if ($query_hapus) {
+if($_SESSION['ses_level'] == "Kadis"){
     echo "<script>
-        Swal.fire({title: 'Hapus Data Berhasil',text: '',icon: 'success',confirmButtonText: 'OK'
-        }).then((result) => {if (result.value) {window.location = 'index.php?page=data-pegawai'
-        ;}})</script>";
-} else {
-    echo "<script>
-        Swal.fire({title: 'Hapus Data Gagal',text: '',icon: 'error',confirmButtonText: 'OK'
-        }).then((result) => {if (result.value) {window.location = 'index.php?page=data-pegawai'
-        ;}})</script>";
+    Swal.fire({title: 'Anda tidak punya akses ke menu ini!',text: '',icon: 'error',confirmButtonText: 'OK'
+    }).then((result) => {if (result.value){
+        window.location = 'index.php?page=data-pegawai';
+        }
+    })</script>";
+}else{
+    $sql_hapus = "DELETE FROM data_pegawai WHERE id_pegawai='" . $_GET['kode'] . "'";
+    $query_hapus = mysqli_query($koneksi, $sql_hapus);
+    if ($query_hapus) {
+        echo "<script>
+            Swal.fire({title: 'Hapus Data Berhasil',text: '',icon: 'success',confirmButtonText: 'OK'
+            }).then((result) => {if (result.value) {window.location = 'index.php?page=data-pegawai'
+            ;}})</script>";
+    } else {
+        echo "<script>
+            Swal.fire({title: 'Hapus Data Gagal',text: '',icon: 'error',confirmButtonText: 'OK'
+            }).then((result) => {if (result.value) {window.location = 'index.php?page=data-pegawai'
+            ;}})</script>";
+    }
 }
